+++ src/dl_txt_unpack.cpp
@@ -90,7 +90,7 @@ static void DLWriteInstance(SDLUnpackContext* _Ctx, const SDLType* _pType, const
 	{
 		const SDLMember& Member = _pType->m_lMembers[i];
 
-		yajl_gen_string(_Ctx->m_JsonGen, (const unsigned char*)Member.m_Name, strlen(Member.m_Name));
+		yajl_gen_string(_Ctx->m_JsonGen, (const unsigned char*)Member.m_Name, (unsigned int)strlen(Member.m_Name));
 
 		const uint8* pMemberData = _pData + Member.m_Offset[DL_PTR_SIZE_HOST];
 
@@ -115,7 +115,7 @@ static void DLWriteInstance(SDLUnpackContext* _Ctx, const SDLType* _pType, const
 					{
 	 					pint  Offset     = *(pint*)pMemberData;
 	 					char* pTheString =  (char*)(_pDataBase + Offset);
-	 					yajl_gen_string(_Ctx->m_JsonGen, (unsigned char*)pTheString, strlen(pTheString));
+	 					yajl_gen_string(_Ctx->m_JsonGen, (unsigned char*)pTheString, (unsigned int)strlen(pTheString));
 	 				}
 					break;
 					case DL_TYPE_STORAGE_PTR:
@@ -136,7 +136,7 @@ static void DLWriteInstance(SDLUnpackContext* _Ctx, const SDLType* _pType, const
 					case DL_TYPE_STORAGE_ENUM:
 					{
 						const char* pEnumName = DLFindEnumName(_Ctx->m_Ctx, Member.m_TypeID, *(uint32*)pMemberData);
-						yajl_gen_string(_Ctx->m_JsonGen, (unsigned char*)pEnumName, strlen(pEnumName));
+						yajl_gen_string(_Ctx->m_JsonGen, (unsigned char*)pEnumName, (unsigned int)strlen(pEnumName));
 					}
 					break;
 					default: // default is a standard pod-type
@@ -175,7 +175,7 @@ static void DLWriteInstance(SDLUnpackContext* _Ctx, const SDLType* _pType, const
 						{
 							pint Offset = *(pint*)(pMemberData + (iElem * Size));
 							char* pStr =   (char*)(_pDataBase + Offset);
-							yajl_gen_string(_Ctx->m_JsonGen, (unsigned char*)pStr, strlen(pStr));
+							yajl_gen_string(_Ctx->m_JsonGen, (unsigned char*)pStr, (unsigned int)strlen(pStr));
 						}
 					}
 					break;
@@ -190,7 +190,7 @@ static void DLWriteInstance(SDLUnpackContext* _Ctx, const SDLType* _pType, const
 							uint32* pEnumData = (uint32*)pMemberData;
 
 							const char* pEnumName = DLFindEnumName(_Ctx->m_Ctx, Member.m_TypeID, pEnumData[iElem]);
-							yajl_gen_string(_Ctx->m_JsonGen, (unsigned char*)pEnumName, strlen(pEnumName));
+							yajl_gen_string(_Ctx->m_JsonGen, (unsigned char*)pEnumName, (unsigned int)strlen(pEnumName));
 						}
 					}
 					break;
@@ -262,7 +262,7 @@ static void DLWriteRoot(SDLUnpackContext* _Ctx, const SDLType* _pType, const uns
 		yajl_gen_map_open(_Ctx->m_JsonGen);
 
 			yajl_gen_string(_Ctx->m_JsonGen, (const unsigned char*)"type", 4);
-			yajl_gen_string(_Ctx->m_JsonGen, (const unsigned char*)_pType->m_Name, strlen(_pType->m_Name));
+			yajl_gen_string(_Ctx->m_JsonGen, (const unsigned char*)_pType->m_Name, (unsigned int)strlen(_pType->m_Name));
 
 			yajl_gen_string(_Ctx->m_JsonGen, (const unsigned char*)"data", 4);
 			DLWriteInstance(_Ctx, _pType, _pData, _pData);
@@ -308,7 +308,7 @@ static void DLWriteRoot(SDLUnpackContext* _Ctx, const SDLType* _pType, const uns
 								{
 									pint  Offset     = *(pint*)(pMemberData + (iElem * sizeof(char*)));
 									char* pTheString =  (char*)(_pData + Offset);
-									yajl_gen_string(_Ctx->m_JsonGen, (unsigned char*)pTheString, strlen(pTheString));
+									yajl_gen_string(_Ctx->m_JsonGen, (unsigned char*)pTheString, (unsigned int)strlen(pTheString));
 								}
 							}
 							break;
@@ -319,7 +319,7 @@ static void DLWriteRoot(SDLUnpackContext* _Ctx, const SDLType* _pType, const uns
 								for(pint iElem = 0; iElem < ArrayCount; ++iElem)
 								{
 									const char* pEnumName = DLFindEnumName(_Ctx->m_Ctx, pMember->m_TypeID, pEnumData[iElem]);
-									yajl_gen_string(_Ctx->m_JsonGen, (unsigned char*)pEnumName, strlen(pEnumName));
+									yajl_gen_string(_Ctx->m_JsonGen, (unsigned char*)pEnumName, (unsigned int)strlen(pEnumName));
 								}
 							}
 							break;
